requirements.txt
h5py
matplotlib
numpy
opencv-python
pandas
Pillow
plotly
psutil
scikit-learn
scipy
seaborn
tensorboard
tensorboard-data-server
tqdm
progress
kaleido
ray[tune]
torch==2.0.1+cu116
torchaudio==2.0.2+cu116
torchsummary
torchvision==0.15.2+cu116
transformers